Increase the speed of the conversion by optimizing the distance metric
The square root is not needed as we don't do a true Euclidian distance but just
want the arg min. The squares could be replaced by abs (Euclidian -> Manhattan)
but then the result becomes blocky. The alpha min beta max seems to suffer from
the same problem. After investigation, it seems to be a rounding error caused by
abs() as introducing abs before squaring and summing reproduce the blockyness